requirements.txt
numpy==1.24.4
torch>=2.0.0
openai-whisper
faster-whisper>=1.0.0
sentence-transformers==2.2.2
transformers>=4.41.0
hdbscan==0.8.33
moviepy>=1.0.0
TTS>=0.6.0
tiktoken
numba
more-itertools
tqdm
python-dotenv
yt-dlp
ffmpeg-python